+++ index.js
@@ -1,9 +1,5 @@
 'use strict';
 
-function onThenSpread(iterable) {
-  return [null, ...iterable];
-}
-
 function onThen(v) {
   return [null, v];
 }
@@ -29,7 +25,7 @@ catchify.race = function catchifyRace(iterable) {
 catchify.all = function catchifyAll(iterable) {
   return Promise
     .all(iterable)
-    .then(onThenSpread, onCatch);
+    .then(onThen, onCatch);
 };
 
 catchify.reject = function catchifyReject(reason){